features: session at database

+++ FastPlaz/src/systems/session_controller.pas
@@ -5,7 +5,7 @@ unit session_controller;
 interface
 
 uses
-  fpcgi, md5, IniFiles,
+  fpcgi, md5, IniFiles, fpjson,
   Classes, SysUtils;
 
 const
@@ -16,11 +16,15 @@ const
   _SESSION_ACTIVE = 'active';         // Start time of session
   _SESSION_KEYSTART = 'start';         // Start time of session
   _SESSION_KEYLAST = 'last';          // Last seen time of session
-  _SESSION_KEYTIMEOUT = 'timeout';       // Timeout in seconds;
   _SESSION_IPADDRESS = 'ipaddr';
   _SESSION_FLASHMESSAGE = 'flash';
   _SESSION_TIMEOUT_DEFAULT = 3600;
+  _SESSION_STORAGE_FILE = 1;
+  _SESSION_STORAGE_DATABASE = 2;
   TDateTimeEpsilon = 2.2204460493e-16;
+
+  _SESSION_SQL_UPDATE =
+    'REPLACE INTO session_info ( sessid, ipaddr, lastused, uid, remember, vars) VALUES( "%s", "%s", now(), %d, %d, "%s");';
 
 type
 
@@ -35,6 +39,7 @@ type
     FSessionPrefix, FSessionSuffix, FSessionExtension, FHttpCookie,
     FCookieID, FSessionID: string;
     FSessionDir: string;
+    FSessionVars: TStringList;
     function GenerateSesionID: string;
     function CreateIniFile(const FileName: string): TMemIniFile;
     procedure DeleteIniFile;
@@ -47,7 +52,17 @@ type
     procedure SetValue(variable: string; AValue: string);
     procedure UpdateIniFile;
 
+    function UpdateDatabase: boolean;
+    function c(SourceString: string): string;
+    function d(SourceString: string): string;
+
+    function StringVarToJson(Source: string): string;
+    function JsonToStringVar(JSON: string): string;
+
+    function StartSessionWithFile: boolean;
+    function StartSessionWithDatabase: boolean;
   public
+    Storage: integer;
     constructor Create();
     destructor Destroy; override;
     property Values[variable: string]: string read GetValue write SetValue; default;
@@ -78,9 +93,12 @@ type
 
 implementation
 
-uses logutil_lib, common;
+uses logutil_lib, common, session_model;
 
 //uses common; --- failed jk memasukkan common ke unit ini
+
+var
+  SessionTable: TSessionModel;
 
 function AppendPathDelim(const Path: string): string;
 begin
@@ -170,17 +188,14 @@ end;
 
 
 function TSessionController.GetIsExpired: boolean;
-var
-  T: integer;
 begin
   Result := False;
-  T := FIniFile.ReadInteger(_SESSION_SESSION, _SESSION_KEYTIMEOUT, FSessionTimeout);
-  if T = 0 then
+  if FSessionTimeout = 0 then
     Exit;
-  if (Interval > T) then
+  if (Interval > FSessionTimeout) then
   begin
+    Clear;
     Result := True;
-    FIniFile.EraseSection(_SESSION_DATA);
   end;
 end;
 
@@ -192,9 +207,18 @@ end;
 function TSessionController.GetValue(variable: string): string;
 begin
   Result := '';
-  if (not FSessionStarted) or (FSessionTerminated) or (FIniFile = nil) then
+  if (not FSessionStarted) or (FSessionTerminated) then
     Exit;
-  Result := FIniFile.ReadString(_SESSION_DATA, variable, '');
+
+  if Storage = _SESSION_STORAGE_FILE then
+  begin
+    if (FIniFile = nil) then
+      Exit;
+    Result := FIniFile.ReadString(_SESSION_DATA, variable, '');
+  end;
+
+  if Storage = _SESSION_STORAGE_DATABASE then
+    Result := FSessionVars.Values[variable];
 end;
 
 procedure TSessionController.SetSessionDir(AValue: string);
@@ -212,18 +236,34 @@ end;
 procedure TSessionController.SetTimeOut(AValue: integer);
 begin
   FSessionTimeout := AValue;
-  FIniFile.WriteInteger(_SESSION_SESSION, _SESSION_KEYTIMEOUT, FSessionTimeout);
 end;
 
 procedure TSessionController.SetValue(variable: string; AValue: string);
 begin
-  if (not FSessionStarted) or (FSessionTerminated) or (FIniFile = nil) then
+  if (not FSessionStarted) or (FSessionTerminated) then
     Exit;
-  try
-    FIniFile.WriteString(_SESSION_DATA, variable, AValue);
-    UpdateIniFile;
-  except
+  if Storage = _SESSION_STORAGE_FILE then
+  begin
+    if (FIniFile = nil) then
+      Exit;
+    try
+      FIniFile.WriteString(_SESSION_DATA, variable, AValue);
+      UpdateIniFile;
+    except
+      on E: Exception do
+      begin
+        pr('session:' + e.Message);
+        die;
+      end;
+    end;
   end;
+
+  if Storage = _SESSION_STORAGE_DATABASE then
+  begin
+    FSessionVars.Values[variable] := AValue;
+    UpdateDatabase;
+  end;
+
 end;
 
 procedure TSessionController.UpdateIniFile;
@@ -257,6 +297,57 @@ begin
   until OK;
 end;
 
+function TSessionController.UpdateDatabase: boolean;
+var
+  sql: string;
+  uid: integer;
+begin
+  uid := s2i(FSessionVars.Values[SESSION_FIELD_UID]);
+  sql := Format(_SESSION_SQL_UPDATE, [FSessionID, Application.Request.RemoteAddress,
+    uid, 0, c(FSessionVars.Text)]);
+  Result := SessionTable.Exec(sql);
+end;
+
+// prepare for encode var session
+function TSessionController.c(SourceString: string): string;
+begin
+  Result := SourceString;
+end;
+
+// prepare for decode var session
+function TSessionController.d(SourceString: string): string;
+begin
+  Result := SourceString;
+end;
+
+function TSessionController.StringVarToJson(Source: string): string;
+var
+  str: TStringList;
+  sessStr: string;
+  i: integer;
+  sessObj: TJSONObject;
+  items: TJSONObject;
+begin
+  sessObj := TJSONObject.Create;
+  items := TJSONObject.Create;
+  str := TStringList.Create;
+  str.Text := Source;
+
+  for i := 0 to str.Count - 1 do
+  begin
+    items.Add(str.Names[i], str.ValueFromIndex[i]);
+  end;
+  sessObj.Add('data', items);
+  Result := (sessObj.AsJSON);
+  FreeAndNil(sessObj);
+  FreeAndNil(str);
+end;
+
+function TSessionController.JsonToStringVar(JSON: string): string;
+begin
+
+end;
+
 function TSessionController._DateTimeDiff(const ANow, AThen: TDateTime): TDateTime;
 begin
   Result := ANow - AThen;
@@ -270,10 +361,18 @@ function TSessionController.GetData: string;
 var
   lst: TStringList;
 begin
-  lst := TStringList.Create;
-  FIniFile.GetStrings(lst);
-  Result := lst.Text;
-  lst.Free;
+  if Storage = _SESSION_STORAGE_FILE then
+  begin
+    lst := TStringList.Create;
+    FIniFile.GetStrings(lst);
+    Result := lst.Text;
+    lst.Free;
+  end;
+
+  if Storage = _SESSION_STORAGE_FILE then
+  begin
+    Result := FSessionVars.Text;
+  end;
 end;
 
 constructor TSessionController.Create;
@@ -281,6 +380,7 @@ var
   lstr: TStrings;
 begin
   inherited Create();
+  FSessionVars := TStringList.Create;
   FLastAccess := 0;
   FHttpCookie := Application.EnvironmentVariable['HTTP_COOKIE'];
   FHttpCookie := StringReplace(FHttpCookie, ' ', '', [rfReplaceAll]);
@@ -293,6 +393,40 @@ begin
   FreeAndNil(lstr);
   FSessionID := GenerateSesionID();
   FSessionDir := Application.EnvironmentVariable['TEMP'];
+  FSessionExtension := '.ses';
+  FSessionStarted := False;
+  FSessionTerminated := False;
+  FCached := False;
+  Storage := _SESSION_STORAGE_FILE;
+  FSessionTimeout := _SESSION_TIMEOUT_DEFAULT;
+end;
+
+destructor TSessionController.Destroy;
+begin
+  inherited Destroy;
+  if Assigned(FIniFile) then
+    FreeAndNil(FIniFile);
+  FreeAndNil(FSessionVars);
+end;
+
+function TSessionController.StartSession: boolean;
+begin
+  Result := False;
+  if FSessionStarted then
+    Exit;
+
+  if Storage = _SESSION_STORAGE_FILE then
+    StartSessionWithFile;
+
+  if Storage = _SESSION_STORAGE_DATABASE then
+    StartSessionWithDatabase;
+end;
+
+function TSessionController.StartSessionWithFile: boolean;
+begin
+  Result := False;
+
+  // check if directory is writeable
   if FSessionDir <> '' then
   begin
     if not DirectoryIsWritable(FSessionDir) then
@@ -308,25 +442,8 @@ begin
     end;
   end;
   FSessionDir := IncludeTrailingPathDelimiter(FSessionDir);
-  FSessionExtension := '.ses';
-  FSessionStarted := False;
-  FSessionTerminated := False;
-  FCached := False;
-  FSessionTimeout := _SESSION_TIMEOUT_DEFAULT;
-end;
 
-destructor TSessionController.Destroy;
-begin
-  inherited Destroy;
-  if Assigned(FIniFile) then
-    FreeAndNil(FIniFile);
-end;
-
-function TSessionController.StartSession: boolean;
-begin
-  Result := False;
-  if FSessionStarted then
-    Exit;
+  //-- create session file
   FIniFile := CreateIniFile(FSessionDir + FSessionID + FSessionExtension);
   if FIniFile = nil then
     Exit;
@@ -336,12 +453,12 @@ begin
     LogUtil.Add('Can''t write session', 'sessions', True);
   end;
 
-  // init session
+  // init session file
   if not FIniFile.ReadBool(_SESSION_SESSION, _SESSION_ACTIVE, False) then
   begin
     FIniFile.WriteBool(_SESSION_SESSION, _SESSION_ACTIVE, True);
-    FIniFile.WriteInteger(_SESSION_SESSION, _SESSION_KEYTIMEOUT, FSessionTimeout);
-    FIniFile.WriteString(_SESSION_SESSION, _SESSION_IPADDRESS, Application.Request.RemoteAddress);
+    FIniFile.WriteString(_SESSION_SESSION, _SESSION_IPADDRESS,
+      Application.Request.RemoteAddress);
     FIniFile.WriteDateTime(_SESSION_SESSION, _SESSION_KEYSTART, now);
     FIniFile.WriteDateTime(_SESSION_SESSION, _SESSION_KEYLAST, now);
   end;
@@ -367,34 +484,87 @@ begin
   Result := True;
 end;
 
+function TSessionController.StartSessionWithDatabase: boolean;
+begin
+  Result := False;
+  SessionTable := TSessionModel.Create();
+  if SessionTable.Find(FSessionID) then
+  begin
+    FSessionVars.Text := d(SessionTable.Value[SESSION_FIELD_VARS]);
+    FLastAccess := SessionTable.Value[SESSION_FIELD_LASTUSED];
+    UpdateDatabase;
+  end
+  else
+  begin
+    FSessionVars.Text := '';
+    UpdateDatabase;
+    if SessionTable.Find(FSessionID) then
+    begin
+      FSessionVars.Text := d(SessionTable.Value[SESSION_FIELD_VARS]);
+      FLastAccess := SessionTable.Value[SESSION_FIELD_LASTUSED];
+    end;
+  end;
+  FSessionStarted := True;
+  Result := True;
+end;
+
 procedure TSessionController.Clear;
 begin
   try
-    FIniFile.EraseSection(_SESSION_DATA);
-    FIniFile.WriteBool(_SESSION_SESSION, _SESSION_ACTIVE, True);
-    FIniFile.WriteInteger(_SESSION_SESSION, _SESSION_KEYTIMEOUT, FSessionTimeout);
-    FIniFile.WriteDateTime(_SESSION_SESSION, _SESSION_KEYSTART, now);
-    FIniFile.WriteDateTime(_SESSION_SESSION, _SESSION_KEYLAST, now);
-    ForceUpdate;
+    if Storage = _SESSION_STORAGE_FILE then
+    begin
+      FIniFile.EraseSection(_SESSION_DATA);
+      FIniFile.WriteBool(_SESSION_SESSION, _SESSION_ACTIVE, True);
+      FIniFile.WriteDateTime(_SESSION_SESSION, _SESSION_KEYSTART, now);
+      FIniFile.WriteDateTime(_SESSION_SESSION, _SESSION_KEYLAST, now);
+      ForceUpdate;
+    end;
+
+    if Storage = _SESSION_STORAGE_DATABASE then
+    begin
+      FSessionVars.Text := '';
+      UpdateDatabase;
+    end;
   except;
   end;
 end;
 
 procedure TSessionController.DeleteKey(const Key: string);
+var
+  i: integer;
 begin
-  FIniFile.DeleteKey(_SESSION_DATA, Key);
+  if Storage = _SESSION_STORAGE_FILE then
+  begin
+    FIniFile.DeleteKey(_SESSION_DATA, Key);
+  end;
+
+  if Storage = _SESSION_STORAGE_DATABASE then
+  begin
+    i := FSessionVars.IndexOfName(Key);
+    if i <> -1 then
+      FSessionVars.Delete(i);
+    UpdateDatabase;
+  end;
 end;
 
 procedure TSessionController.EndSession(const Force: boolean);
 begin
   try
-    FIniFile.WriteBool(_SESSION_SESSION, _SESSION_ACTIVE, False);
-    FIniFile.EraseSection(_SESSION_DATA);
-    if Force then
+    if Storage = _SESSION_STORAGE_FILE then
     begin
-      DeleteIniFile;
-      FreeAndNil(FIniFile);
-      FSessionTerminated := True;
+      FIniFile.WriteBool(_SESSION_SESSION, _SESSION_ACTIVE, False);
+      FIniFile.EraseSection(_SESSION_DATA);
+      if Force then
+      begin
+        DeleteIniFile;
+        FreeAndNil(FIniFile);
+        FSessionTerminated := True;
+      end;
+    end;
+
+    if Storage = _SESSION_STORAGE_DATABASE then
+    begin
+      Clear;
     end;
   except
   end;
@@ -407,23 +577,47 @@ end;
 
 procedure TSessionController.ForceUpdate;
 begin
-  UpdateIniFile;
+  if Storage = _SESSION_STORAGE_FILE then
+    UpdateIniFile;
+  if Storage = _SESSION_STORAGE_DATABASE then
+    UpdateDatabase;
 end;
 
 function TSessionController.ReadDateTime(const variable: string): TDateTime;
 begin
   Result := 0;
-  if (not FSessionStarted) or (FSessionTerminated) or (FIniFile = nil) then
-    Exit;
-  Result := FIniFile.ReadDateTime(_SESSION_DATA, variable, 0);
+
+  if Storage = _SESSION_STORAGE_FILE then
+  begin
+    if (not FSessionStarted) or (FSessionTerminated) or (FIniFile = nil) then
+      Exit;
+    Result := FIniFile.ReadDateTime(_SESSION_DATA, variable, 0);
+  end;
+
+  if Storage = _SESSION_STORAGE_DATABASE then
+  begin
+    try
+      Result := StrToDateTime(FSessionVars.Values[variable]);
+    except
+    end;
+  end;
 end;
 
 function TSessionController.ReadInteger(const variable: string): integer;
 begin
   Result := 0;
-  if (not FSessionStarted) or (FSessionTerminated) or (FIniFile = nil) then
-    Exit;
-  Result := FIniFile.ReadInteger(_SESSION_DATA, variable, 0);
+
+  if Storage = _SESSION_STORAGE_FILE then
+  begin
+    if (not FSessionStarted) or (FSessionTerminated) or (FIniFile = nil) then
+      Exit;
+    Result := FIniFile.ReadInteger(_SESSION_DATA, variable, 0);
+  end;
+
+  if Storage = _SESSION_STORAGE_DATABASE then
+  begin
+    Result := s2i(FSessionVars.Values[variable]);
+  end;
 end;
 
 end.
